requirements.txt
# Archive Tools Project - Requirements
# المكتبات المطلوبة لمشروع أدوات الأرشيف

# ============ المكتبات الأساسية ============
# ملاحظة: zipfile, tarfile, gzip, sqlite3, pathlib, re, json, csv
# مضمنة في بايثون القياسية

# ============ دعم الأرشيفات ============
rarfile>=4.0              # دعم ملفات RAR

# ============ معالجة PDF ============
pdfplumber>=0.9.0         # استخراج النصوص والجداول من PDF
Pillow>=10.0.0            # معالجة الصور
pytesseract>=0.3.10       # OCR - التعرف الضوئي على الحروف

# ============ معالجة المستندات ============
python-docx>=0.8.11       # قراءة ملفات Word
openpyxl>=3.1.2           # قراءة ملفات Excel
pandas>=2.0.0             # معالجة البيانات والجداول

# ============ معالجة HTML ============
beautifulsoup4>=4.12.0    # تحليل HTML
lxml>=4.9.0               # محرك تحليل سريع

# ============ أدوات مساعدة ============
chardet>=5.2.0            # كشف ترميز الملفات
tqdm>=4.66.1              # شريط التقدم

# ============ مكتبات اختيارية ============
# datasketch>=1.6.0       # MinHash LSH لإزالة التكرارات
# scikit-learn>=1.3.0     # التصنيف والتجميع
# numpy>=1.24.0           # العمليات الحسابية
